Plug query_lock_limiter into QueryCountingApiClient.
